Fitness Function Improved and SWT Dependency added
A redesigned Fitness Function calculator was added. It takes into
account the accuracy of the decision tree while classifying both,
training and test examples. The fitness score is a function of the
weighted average of the two accuracies.

A small window was created for testing SWT. Run the MainWindow class and
click the "Click Me" button when the window opens.