Decision Tree Classifier GUI Updated
The updated GUI now displays sliders that the user can use to vary the
Crossover and Mutation Probabilities. The samples from the selected
dataset are displayed in a Table Widget. After the Genetic Algorithm has
completed its iterations, the result is displayed in a small text area.
Also, the misclassified examples are displayed in "red" color in the
table.